+++ test/history_test.js
@@ -351,7 +351,6 @@ describe('History', function() {
         browser.document.body.innerHTML = '<html><body>Wolf</body></html>';
         browser.window.addEventListener('hashchange', ()=> done());
         browser.window.location.hash = 'boo';
-        browser.wait();
       });
 
       it('should add page to history', function() {